[FIX] pace int로 변경

See merge request s08-bigdata-recom-sub2/S08P22D204!111

+++ src/main/java/com/d204/rumeet/record/model/dto/RaceModeInfoDto.java
@@ -9,7 +9,7 @@ public class RaceModeInfoDto {
     int mode;
     long time;
     float km;
-    float pace;
+    int pace;
     int heartRate;
     double kcal;
     int success;
